requirements.txt
appdirs==1.4.4
apturl==0.5.2
asn1crypto==0.24.0
bcrypt==3.1.7
beautifulsoup4==4.9.0
blinker==1.4
Brlapi==0.6.6
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distlib==0.3.0
distro-info===0.18ubuntu0.18.04.1
dnspython==1.16.0
email-validator==1.1.0
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
html5lib==1.0.1
httplib2==0.9.2
idna==2.9
importlib-metadata==1.6.0
importlib-resources==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==7.1.2
protobuf==3.0.0
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.23.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.17
system-service==0.3
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.9
usb-creator==0.3.3
virtualenv==20.0.20
wadllib==1.3.2
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.1
xkit==0.0.0
zipp==3.1.0
zope.interface==4.3.2
